+++ cmd/root.go
@@ -5,25 +5,17 @@ import (
 	"fmt"
 	"os"
 
+	cliconfigsetup "github.com/aideneyre/acli/cmd/config"
 	help "github.com/aideneyre/acli/cmd/help"
 	kube "github.com/aideneyre/acli/cmd/kube"
-	"github.com/aideneyre/acli/internal/config"
 
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
-
-var AcliConfig viper.Viper
 
 var RootCmd = &cobra.Command{
 	Use:     "acli [flags] [options]",
 	Short:   "Aiden CLI - quality of life commands",
 	Version: "1.3.0",
-
-	PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
-		err := config.Initialize()
-		return err
-	},
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Printf("you are using acli version %s\n\n", cmd.Version)
 		cmd.Help()
@@ -35,6 +27,7 @@ var RootCmd = &cobra.Command{
 func init() {
 	RootCmd.AddCommand(kube.KubeChangeCmd)
 	RootCmd.AddCommand(help.CheatSheet)
+	RootCmd.AddCommand(cliconfigsetup.ConfigCmd)
 }
 
 // Execute runs the command and its subcommands, returning an error